+++ src/game/effects/base/requests.cpp
@@ -168,7 +168,7 @@ namespace banggame {
     }
 
     game_formatted_string request_discard_pass::status_text(player *owner) const {
-        int diff = target->m_hand.size() - target->max_cards_end_of_turn();
+        int diff = static_cast<int>(target->m_hand.size() - target->max_cards_end_of_turn());
         if (diff > 1) {
             if (target == owner) {
                 return {"STATUS_DISCARD_PASS_PLURAL", diff};
