+++ src/game/request_queue.h
@@ -79,9 +79,9 @@ namespace banggame {
 
         template<std::invocable Function>
         int num_queued_requests(Function &&fun) {
-            int nreqs = m_requests.size();
+            size_t nreqs = m_requests.size();
             std::invoke(std::forward<Function>(fun));
-            return m_requests.size() - nreqs;
+            return static_cast<int>(m_requests.size() - nreqs);
         }
 
         void tick() {
